Allow for measuring the source up-to-dateness of Harbor JSON reports.
Closes #10609.
