+++ components/collector/.vulture_ignore_list.py
@@ -68,7 +68,9 @@ GitLabSourceVersion  # unused class (src/source_collectors/gitlab/source_version
 GitLabUnusedJobs  # unused class (src/source_collectors/gitlab/unused_jobs.py:11)
 scan_status  # unused variable (src/source_collectors/harbor/security_warnings.py:58)
 HarborSecurityWarnings  # unused class (src/source_collectors/harbor/security_warnings.py:62)
-HarborJSONSecurityWarnings  # unused class (src/source_collectors/harbor_json/security_warnings.py:36)
+generated_at  # unused variable (src/source_collectors/harbor_json/base.py:23)
+HarborJSONSecurityWarnings  # unused class (src/source_collectors/harbor_json/security_warnings.py:12)
+HarborJSONSourceUpToDateness  # unused class (src/source_collectors/harbor_json/source_up_to_dateness.py:13)
 JacocoSourceUpToDateness  # unused class (src/source_collectors/jacoco/source_up_to_dateness.py:11)
 JacocoUncoveredBranches  # unused class (src/source_collectors/jacoco/uncovered_branches.py:6)
 JacocoUncoveredLines  # unused class (src/source_collectors/jacoco/uncovered_lines.py:6)
