+++ components/collector/src/source_collectors/__init__.py
@@ -60,6 +60,7 @@ from .gitlab.source_version import GitLabSourceVersion
 from .gitlab.unused_jobs import GitLabUnusedJobs
 from .harbor.security_warnings import HarborSecurityWarnings
 from .harbor_json.security_warnings import HarborJSONSecurityWarnings
+from .harbor_json.source_up_to_dateness import HarborJSONSourceUpToDateness
 from .jacoco.source_up_to_dateness import JacocoSourceUpToDateness
 from .jacoco.uncovered_branches import JacocoUncoveredBranches
 from .jacoco.uncovered_lines import JacocoUncoveredLines
